requirements.txt
flask
google-generativeai
requests
selenium